Adding code to deal with failure of github.  Luckily github had a failed db maintenance while i was working on this and noticed the problem. To enable, copy setec-astronomy.zip to s3 and set url appropriatly.
If github is down the code is never downloaded and the server just sits there without web port being open.

In the end we should be doing direct ssh connections for manipulating the server rather than relying on a web app.  At that point we could connect via ssh and wait for the gems and whatnot to be installed then activate services and copy files.  It won't be faster but it would be more "secure".  A web interface for manipulating things will be handy but for automation it's not really neccessary.

diff --git a/aws/setec-astronomy.template b/aws/setec-astronomy.template
index 9524d1a..df3c8f5 100755
--- a/aws/setec-astronomy.template
+++ b/aws/setec-astronomy.template
@@ -51,6 +51,8 @@
       "gem install sinatra","\n",
       "# pull down source tree, extract, and copy files to and fro","\n",
       "wget https://github.com/mshirley/setec-astronomy/zipball/master -O /tmp/setec-astronomy.zip","\n",
+      "CODE=$?","\n",
+      "if [ $CODE -ne 0 ]; then echo 'failed to download source, falling back to to s3'; wget https://s3bucketurl-changeme; fi","\n",
       "unzip /tmp/setec-astronomy.zip -d /tmp","\n",
       "mv /tmp/mshirley-setec-* /tmp/setec-astronomy","\n",
       "cp /tmp/setec-astronomy/conf/named/named.* /etc/","\n",